+++ bindings/nostr-sdk-js/src/protocol/types/time.rs
@@ -53,6 +53,6 @@ impl JsTimestamp {
     /// Convert `Timestamp` to human datetime
     #[wasm_bindgen(js_name = toHumanDatetime)]
     pub fn to_human_datetime(&self) -> String {
-        self.inner.to_human_datetime().into_owned()
+        self.inner.to_human_datetime()
     }
 }