+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 * nostr: change `EventBuilder::gift_wrap` (and linked methods) args to take `extra_tags` instead of `expiration` ([erskingardner])
 * nostr: disable all default features except `std` ([Yuki Kishimoto])
+* nostr: change `Timestamp::to_human_datetime` fingerprint ([Yuki Kishimoto])
 * pool: switch from async to sync message sending for `Relay` ([Yuki Kishimoto])
 * sdk: disable all default features ([Yuki Kishimoto])
 * sdk: set `Client::from_builder` as private ([Yuki Kishimoto])